==> 00_logistic_models.R <==
agesex_forms <- list(bin.exp ~ Age + Sex + (1|RegionalID),
                     bin.exp ~ Age * Sex + (1|RegionalID),
                     bin.exp ~ Sex + (1|RegionalID),
                     bin.exp ~ Age + (1|RegionalID),
                     bin.exp ~ Sex + Age + I(Age^2) + (1|RegionalID))
names(agesex_forms) <- as.character(agesex_forms)

ag_formulae <- list(bin.exp ~ totalag_15 + (1|RegionalID),
                    bin.exp ~ totalag_30 + (1|RegionalID),
                    bin.exp ~ totalag_60 + (1|RegionalID),
                    bin.exp ~ crops_15 + (1|RegionalID),
                    bin.exp ~ crops_30 + (1|RegionalID),
                    bin.exp ~ crops_60 + (1|RegionalID),
                    bin.exp ~ pasture_15 + (1|RegionalID),
                    bin.exp ~ pasture_30 + (1|RegionalID),
                    bin.exp ~ pasture_60 + (1|RegionalID))
names(ag_formulae) <- as.character(ag_formulae)

wui_formulae <- list(bin.exp ~ mix_15_100 + (1|RegionalID),
                     bin.exp ~ mix_30_100 + (1|RegionalID),
                     bin.exp ~ mix_60_100 + (1|RegionalID),
                     bin.exp ~ mix_15_250 + (1|RegionalID),
                     bin.exp ~ mix_30_250 + (1|RegionalID),
                     bin.exp ~ mix_60_250 + (1|RegionalID),
                     bin.exp ~ mix_15_500 + (1|RegionalID),
                     bin.exp ~ mix_30_500 + (1|RegionalID),
                     bin.exp ~ mix_60_500 + (1|RegionalID),
                     bin.exp ~ face_15_100 + (1|RegionalID),
                     bin.exp ~ face_30_100 + (1|RegionalID),
                     bin.exp ~ face_60_100 + (1|RegionalID),
                     bin.exp ~ face_15_250 + (1|RegionalID),
                     bin.exp ~ face_30_250 + (1|RegionalID),
                     bin.exp ~ face_60_250 + (1|RegionalID),
                     bin.exp ~ face_15_500 + (1|RegionalID),
                     bin.exp ~ face_30_500 + (1|RegionalID),
                     bin.exp ~ face_60_500 + (1|RegionalID),
                     bin.exp ~ wui_15_100 + (1|RegionalID),
                     bin.exp ~ wui_30_100 + (1|RegionalID),
                     bin.exp ~ wui_60_100 + (1|RegionalID),
                     bin.exp ~ wui_15_250 + (1|RegionalID),
                     bin.exp ~ wui_30_250 + (1|RegionalID),
                     bin.exp ~ wui_60_250 + (1|RegionalID),
                     bin.exp ~ wui_15_500 + (1|RegionalID),
                     bin.exp ~ wui_30_500 + (1|RegionalID),
                     bin.exp ~ wui_60_500 + (1|RegionalID))
names(wui_formulae) <- as.character(wui_formulae)

beech_formulae <- list(bin.exp ~ BBA_15 + (1|RegionalID),
                       bin.exp ~ BBA_30 + (1|RegionalID),
                       bin.exp ~ BBA_60 + (1|RegionalID))
names(beech_formulae) <- as.character(beech_formulae)


forest_formulae <- list(bin.exp ~ totalforest_15 + (1|RegionalID),
                        bin.exp ~ totalforest_30 + (1|RegionalID),
                        bin.exp ~ totalforest_60 + (1|RegionalID),
                        bin.exp ~ deciduous_15 + (1|RegionalID),
                        bin.exp ~ deciduous_30 + (1|RegionalID),
                        bin.exp ~ deciduous_60 + (1|RegionalID),
                        bin.exp ~ evergreen_15 + (1|RegionalID),
                        bin.exp ~ evergreen_30 + (1|RegionalID),
                        bin.exp ~ evergreen_60 + (1|RegionalID),
                        bin.exp ~ mixed_15 + (1|RegionalID),
                        bin.exp ~ mixed_30 + (1|RegionalID),
                        bin.exp ~ mixed_60 + (1|RegionalID))
names(forest_formulae) <- as.character(forest_formulae)

build_formulae <- list(bin.exp ~ nbuildings_15 + (1|RegionalID),
                       bin.exp ~ nbuildings_30 + (1|RegionalID),
                       bin.exp ~ nbuildings_60 + (1|RegionalID),
                       bin.exp ~ build_cat_15 + (1|RegionalID),
                       bin.exp ~ build_cat_30 + (1|RegionalID),
                       bin.exp ~ build_cat_60 + (1|RegionalID))
names(build_formulae) <- as.character(build_formulae)


# Full models for 

brod_models <- list(bin.exp ~ Sex + Age + build_cat_15*totalforest_15 + totalag_15 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_30*totalforest_30 + totalag_30 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_60*totalforest_60 + totalag_60 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_15*totalforest_15 + crops_15 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_30*totalforest_30 + crops_30 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_60*totalforest_60 + crops_60 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_15*totalforest_15 + pasture_15 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_30*totalforest_30 + pasture_30 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_60*totalforest_60 + pasture_60 + lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_15*totalforest_15 + totalag_15 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_30*totalforest_30 + totalag_30 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_60*totalforest_60 + totalag_60 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_15*totalforest_15 + crops_15 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_30*totalforest_30 + crops_30 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_60*totalforest_60 + crops_60 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_15*totalforest_15 + pasture_15 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_30*totalforest_30 + pasture_30 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + build_cat_60*totalforest_60 + pasture_60 + beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_15_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_15_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_15_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID), 
                    bin.exp ~ Sex + Age + wui_15_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_15_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_15_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_15_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_15_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_15_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_30_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + wui_60_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID), 
                    bin.exp ~ Sex + Age + mix_15_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_15_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_15_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID), 
                    bin.exp ~ Sex + Age + mix_15_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_15_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_15_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID), 
                    bin.exp ~ Sex + Age + mix_15_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_15_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_15_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_30_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + mix_60_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID), 
                    bin.exp ~ Sex + Age + face_15_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_15_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_15_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + totalag_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + totalag_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + totalag_60 + BBA_60 * lag_beechnuts + (1|RegionalID), 
                    bin.exp ~ Sex + Age + face_15_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_15_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_15_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + pasture_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + pasture_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + pasture_60 + BBA_60 * lag_beechnuts + (1|RegionalID), 
                    bin.exp ~ Sex + Age + face_15_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_15_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_15_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_30_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + crops_15 + BBA_15 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + crops_30 + BBA_30 * lag_beechnuts + (1|RegionalID),
                    bin.exp ~ Sex + Age + face_60_100 + crops_60 + BBA_60 * lag_beechnuts + (1|RegionalID))
names(brod_models) <- as.character(brod_models)
